requirements.txt
# Core ML and NLP
torch>=2.0.0
transformers>=4.30.0
sentence-transformers>=2.2.0
accelerate>=0.20.0

# Document processing
PyPDF2>=3.0.0
python-docx>=0.8.11
docx2txt>=0.8

# Web framework and API
fastapi>=0.100.0
uvicorn>=0.22.0
python-multipart>=0.0.6

# Optional: OpenAI for higher quality summaries
openai>=1.0.0

# Utilities
numpy>=1.24.0
pandas>=2.0.0
tqdm>=4.65.0
python-dotenv>=1.0.0

# Development
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0 